# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Aiexec
  version: 1.6.4
paths:
...
  /api/v2/files/:
    get:
      tags:
      - Files
      summary: List Files
      description: List the files available to the current user.
      operationId: list_files_api_v2_files__get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/File'
                type: array
                title: Response List Files Api V2 Files  Get
      security:
      - OAuth2PasswordBearer: []
      - API key query: []
      - API key header: []
    post:
      tags:
      - Files
      summary: Upload User File
      description: Upload a file for the current user and track it in the database.
      operationId: upload_user_file_api_v2_files__post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_user_file_api_v2_files__post'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/aiexec__api__schemas__UploadFileResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
      - API key query: []
      - API key header: []
    delete:
      tags:
      - Files
      summary: Delete All Files
      description: Delete all files for the current user.
      operationId: delete_all_files_api_v2_files__delete
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - OAuth2PasswordBearer: []
      - API key query: []
      - API key header: []
  /api/v2/files:
    get:
      tags:
      - Files
      summary: List Files
      description: List the files available to the current user.
      operationId: list_files_api_v2_files_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/File'
                type: array
                title: Response List Files Api V2 Files Get
      security:
      - OAuth2PasswordBearer: []
      - API key query: []
      - API key header: []
    post:
      tags:
      - Files
      summary: Upload User File
      description: Upload a file for the current user and track it in the database.
      operationId: upload_user_file_api_v2_files_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_user_file_api_v2_files_post'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/aiexec__api__schemas__UploadFileResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
      - API key query: []
      - API key header: []
    delete:
      tags:
      - Files
      summary: Delete All Files
      description: Delete all files for the current user.
      operationId: delete_all_files_api_v2_files_delete
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - OAuth2PasswordBearer: []
      - API key query: []
      - API key header: []
  /api/v2/files/batch/:
    post:
      tags:
      - Files
      summary: Download Files Batch
      description: Download multiple files as a zip file by their IDs.
      operationId: download_files_batch_api_v2_files_batch__post
      requestBody:
        content:
          application/json:
            schema:
              items:
                type: string
                format: uuid
              type: array
              title: File Ids
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
      - API key query: []
      - API key header: []
    delete:
      tags:
      - Files
      summary: Delete Files Batch
      description: Delete multiple files by their IDs.
      operationId: delete_files_batch_api_v2_files_batch__delete
      requestBody:
        content:
          application/json:
            schema:
              items:
                type: string
                format: uuid
              type: array
              title: File Ids
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
      - API key query: []
      - API key header: []
  /api/v2/files/{file_id}:
    get:
      tags:
      - Files
      summary: Download File
      description: "Download a file by its ID or return its content as a string/bytes.\n\
        \nArgs:\n    file_id: UUID of the file.\n    current_user: Authenticated user.\n\
        \    session: Database session.\n    storage_service: File storage service.\n\
        \    return_content: If True, return raw content (str) instead of StreamingResponse.\n\
        \nReturns:\n    StreamingResponse for client downloads or str for internal\
        \ use."
      operationId: download_file_api_v2_files__file_id__get
      security:
      - OAuth2PasswordBearer: []
      - API key query: []
      - API key header: []
      parameters:
      - name: file_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: File Id
      - name: return_content
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Return Content
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - Files
      summary: Edit File Name
      description: Edit the name of a file by its ID.
      operationId: edit_file_name_api_v2_files__file_id__put
      security:
      - OAuth2PasswordBearer: []
      - API key query: []
      - API key header: []
      parameters:
      - name: file_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: File Id
      - name: name
        in: query
        required: true
        schema:
          type: string
          title: Name
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/aiexec__api__schemas__UploadFileResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Files
      summary: Delete File
      description: Delete a file by its ID.
      operationId: delete_file_api_v2_files__file_id__delete
      security:
      - OAuth2PasswordBearer: []
      - API key query: []
      - API key header: []
      parameters:
      - name: file_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: File Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    Body_upload_user_file_api_v2_files__post:
      properties:
        file:
          type: string
          format: binary
          title: File
      type: object
      required:
      - file
      title: Body_upload_user_file_api_v2_files__post
    Body_upload_user_file_api_v2_files_post:
      properties:
        file:
          type: string
          format: binary
          title: File
      type: object
      required:
      - file
      title: Body_upload_user_file_api_v2_files_post
...
    File:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        user_id:
          type: string
          format: uuid
          title: User Id
        name:
          type: string
          title: Name
        path:
          type: string
          title: Path
        size:
          type: integer
          title: Size
        provider:
          anyOf:
          - type: string
          - type: 'null'
          title: Provider
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - user_id
      - name
      - path
      - size
      title: File
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    aiexec__api__schemas__UploadFileResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        name:
          type: string
          title: Name
        path:
          type: string
          format: path
          title: Path
        size:
          type: integer
          title: Size
        provider:
          anyOf:
          - type: string
          - type: 'null'
          title: Provider
      type: object
      required:
      - id
      - name
      - path
      - size
      title: UploadFileResponse
      description: File upload response schema.
...
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: api/v1/login
    API key query:
      type: apiKey
      in: query
      name: x-api-key
    API key header:
      type: apiKey
      in: header
      name: x-api-key